fix(auth): update user seeding logic and enhance JWT session verification
